requirements.txt
requests
praw
logging
python-decouple